Keep the repository constant within session
Hi Mike, 
I have added a couple of lines to keep the Repository constant in two conditions: 
1. Multiple runs of GET all autos(). 
If the content keeps the same, it would be easier for those who are playing with the API code to track the changes. The current behavior is that when all_cars() is called it's reloading the dict with new uuids. 
2. Keep the API behavior the same in responding to different sequence of requests. 
In your video, you always get the list of all cars before you try out any other requests. However, what if someone is tweaking the code, restarts the server, adds a new car before listing all cars? When I was using postman testing the code, it happened to me and I was quite surprised to see there's only one car returned from the API. The sequence of the request will cause the API to behave differently. 

I know this code is just for demo purpose. But the experience might be smoother if the the API behaves consistently when someone is playing with the code. What do you think?